add all runtime tests
